services/processing/requirements.txt
kafka-python-ng==2.2.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
numpy==1.26.3
python-dotenv==1.0.0
